+++ src/renderer/src/modules/discover/recommendations.tsx
