+++ src/renderer/src/modules/entry-column/social-media-item.tsx
@@ -4,7 +4,7 @@ import dayjs from "@renderer/lib/dayjs"
 import { cn } from "@renderer/lib/utils"
 import { useEntry } from "@renderer/store/entry/hooks"
 
-import { ReactVirtuosoItemPlaceholder } from "../ui/placeholder"
+import { ReactVirtuosoItemPlaceholder } from "../../components/ui/placeholder"
 import { EntryTranslation } from "./translation"
 import type { UniversalItemProps } from "./types"
