+++ src/renderer/src/modules/entry-column/translation.tsx
