+++ src/renderer/src/modules/entry-column/types.ts
