+++ src/renderer/src/modules/entry-content/share.tsx
