+++ src/renderer/src/modules/feed-column/index.tsx
@@ -10,7 +10,7 @@ import { Lethargy } from "lethargy"
 import { useEffect, useRef, useState } from "react"
 import { Link } from "react-router-dom"
 
-import { Vibrancy } from "../ui/background"
+import { Vibrancy } from "../../components/ui/background"
 import { FeedList } from "./list"
 
 const lethargy = new Lethargy()
