+++ src/renderer/src/modules/settings/action-card.tsx
