+++ src/renderer/src/modules/settings/title.tsx
