+++ src/renderer/src/pages/(external)/(with-layout)/feed/[:id]/index.tsx
@@ -1,9 +1,3 @@
-import { ArticleItem } from "@renderer/components/entry-column/article-item"
-import { NotificationItem } from "@renderer/components/entry-column/notification-item"
-import { PictureItem } from "@renderer/components/entry-column/picture-item"
-import { SocialMediaItem } from "@renderer/components/entry-column/social-media-item"
-import type { UniversalItemProps } from "@renderer/components/entry-column/types"
-import { VideoItem } from "@renderer/components/entry-column/video-item"
 import { FeedIcon } from "@renderer/components/feed-icon"
 import { FollowIcon } from "@renderer/components/icons/follow"
 import { StyledButton } from "@renderer/components/ui/button"
@@ -12,6 +6,12 @@ import { APP_NAME, views } from "@renderer/lib/constants"
 import { FeedViewType } from "@renderer/lib/enum"
 import { cn } from "@renderer/lib/utils"
 import type { FeedModel } from "@renderer/models"
+import { ArticleItem } from "@renderer/modules/entry-column/article-item"
+import { NotificationItem } from "@renderer/modules/entry-column/notification-item"
+import { PictureItem } from "@renderer/modules/entry-column/picture-item"
+import { SocialMediaItem } from "@renderer/modules/entry-column/social-media-item"
+import type { UniversalItemProps } from "@renderer/modules/entry-column/types"
+import { VideoItem } from "@renderer/modules/entry-column/video-item"
 import { useEntriesPreview } from "@renderer/queries/entries"
 import { useFeed } from "@renderer/queries/feed"
 import { DEEPLINK_SCHEME } from "@shared/constants"
