+++ src/renderer/src/pages/settings/actions.tsx
@@ -1,9 +1,9 @@
-import { ActionCard } from "@renderer/components/settings/action-card"
-import { SettingsTitle } from "@renderer/components/settings/title"
 import { Button } from "@renderer/components/ui/button"
 import { useBizQuery } from "@renderer/hooks"
 import { apiClient } from "@renderer/lib/api-fetch"
 import type { ActionsResponse } from "@renderer/models"
+import { ActionCard } from "@renderer/modules/settings/action-card"
+import { SettingsTitle } from "@renderer/modules/settings/title"
 import { Queries } from "@renderer/queries"
 import { useMutation } from "@tanstack/react-query"
 import { useEffect, useState } from "react"
