+++ src/renderer/src/pages/settings/index.tsx
@@ -1,5 +1,5 @@
 import { AppearanceSwitch } from "@renderer/components/appearance-switch"
-import { SettingsTitle } from "@renderer/components/settings/title"
+import { SettingsTitle } from "@renderer/modules/settings/title"
 
 export function Component() {
   return (
